[ET-VK][AOT][ez] Introduce vulkan export utils lib
## Changes

As title. Introduce a common Python utility library for scripts in the Vulkan backend.

Differential Revision: [D65291064](https://our.internmc.facebook.com/intern/diff/D65291064/)

[ghstack-poisoned]
